Feat : 레벨형 배지 계산 클래스 이름 변경
