+++ src/main/java/noodlezip/badge/service/category/AllCommunityPostCountBadge.java
@@ -2,14 +2,14 @@ package noodlezip.badge.service.category;
 
 import lombok.RequiredArgsConstructor;
 import noodlezip.badge.constants.LevelBadgeCategoryType;
-import noodlezip.badge.service.process.LevelStrategyDirectUpdateProcessor;
+import noodlezip.badge.service.process.LevelDirectUpdateProcessor;
 import org.springframework.stereotype.Component;
 
 @RequiredArgsConstructor
 @Component
 public class AllCommunityPostCountBadge {
 
-    private final LevelStrategyDirectUpdateProcessor directUpdateProcessor;
+    private final LevelDirectUpdateProcessor directUpdateProcessor;
 
     public void process(Long userId) {
         directUpdateProcessor.process(userId, LevelBadgeCategoryType.ALL_COMMENT_POST_COUNT_BADGE);
