+++ src/main/java/noodlezip/badge/service/event/review/RamenReviewLikeEvent.java
@@ -4,14 +4,14 @@ import lombok.RequiredArgsConstructor;
 import noodlezip.badge.constants.LevelBadgeCategoryType;
 import noodlezip.badge.dto.EmptyInfoDto;
 import noodlezip.badge.service.event.BadgeEventReader;
-import noodlezip.badge.service.process.LevelStrategyDirectUpdateProcessor;
+import noodlezip.badge.service.process.LevelDirectUpdateProcessor;
 import org.springframework.stereotype.Component;
 
 @RequiredArgsConstructor
 @Component
 public class RamenReviewLikeEvent implements BadgeEventReader<EmptyInfoDto> {
 
-    private final LevelStrategyDirectUpdateProcessor directUpdateProcessor;
+    private final LevelDirectUpdateProcessor directUpdateProcessor;
 
     @Override
     public void read(Long userId, EmptyInfoDto extraOption) {
